+++ src/scene.cpp
@@ -4,10 +4,10 @@
 
 namespace rt {
 
-bool Scene::Intersect(const Ray &ray, const float t_min, const float t_max, HitRecord &rec) const noexcept
+bool Scene::Intersect(const Ray &ray, const double t_min, const double t_max, HitRecord &rec) const noexcept
 {
   bool hit_anything = false;
-  float closest_so_far = t_max;
+  double closest_so_far = t_max;
 
   std::for_each(std::execution::seq, _objects.begin(), _objects.end(), [&](const auto &object) {
     if (object->Intersect(ray, t_min, closest_so_far, rec)) {
